+++ ubuntu/18.04/php/install.sh
@@ -32,6 +32,14 @@ source "${ABSPKG}/functions.sh"
 echo
 echo "Start installing ${PKGNAME^^}."
 
+# Make sure the package is installed.
+if [ ! -z "$(is${PKGNAME^})" ]; then
+  CONFIRM_MESSAGE="$(msg -yn "The ${PKGNAME} package is already installed. Would you like to reinstall?")"
+  if [ CONFIRM_MESSAGE == "No" ]; then
+    exit 0
+  fi
+fi
+
 # Installing php extensions for amp.
 apt -y install php php-common libapache2-mod-php php-mysql
